now finds minimum height until negative normal force
added a while loop that increments H as long as there is no negative normal force. This now tells us the minimum height until we reach a height in the loop that creates a negative normal force.

This is not a finished product. The first appearance of negative normal is not when our block completes the loop. Currently the block makes it partway up, then stalls back and forth.

Don't exactly know how to fix, but I'm  pretty sure I'm on the right track here. Some other condition is needed in the while loop.

Something like when the fNorm at 90 degrees = fNorm at 180 degrees